requirements.txt
python-telegram-bot==20.6
python-dotenv==1.0.0
asyncpg==0.28.0
openai==0.28.1
cachetools==5.3.1
sentry-sdk==1.31.0